Add files via upload
This application simulates the passive diffusion of small molecule drugs from a cyclodextrin-based hydrogel from a variety of directions in a variety of ordinary geometries including slabs, cylinders, prisms, and spheres. By taking into account several experimentally derived parameters such as association/disassociation rate constants, concentration, and the dimensions of the hydrogel, this simulations predicts the release profile of the drug over time for the design of smart therapeutics.